django-backend/requirements.txt
angular-django==0.0.1
asgiref==3.8.1
black==24.4.2
click==8.1.7
Django==5.0.6
django-cors-headers==4.4.0
django-phonenumber-field==7.3.0
djangorestframework==3.15.1
mypy-extensions==1.0.0
mysql-connector-python==8.4.0
packaging==24.1
pathspec==0.12.1
phonenumbers==8.13.39
platformdirs==4.2.2
sqlparse==0.5.0
